Fixing custom markdown in code blocks and fend multi-paragraph spans
