+++ app/src/app/types/index.tsx
@@ -53,7 +53,7 @@ declare global {
 export const APP_NAME = 'µPad';
 export const SYNC_NAME = 'µSync';
 export const MICROPAD_URL = 'https://getmicropad.com';
-export const UNSUPPORTED_MESSAGE = 'Support for this type of content was removed in v3. You can go to https://getmicropad.com/web to access v2.';
+export const UNSUPPORTED_MESSAGE = 'Support for this type of content has been removed. You can go to https://getmicropad.com/web to access v2.';
 export const BAD_BROWSER_AUDIO = `If your web browser doesn't support this type of audio you can click here to download it.`;
 
 // Help messages
